feat: 기능이 아닌 도메인에 따라 패키지 구조 변경
