+++ backend/src/main/java/softeer/wantcar/cartalog/trim/dto/TrimExteriorColorListResponseDto.java
@@ -1,4 +1,4 @@
-package softeer.wantcar.cartalog.dto;
+package softeer.wantcar.cartalog.trim.dto;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
