+++ backend/src/test/java/softeer/wantcar/cartalog/controller/TrimColorControllerTest.java
@@ -7,12 +7,13 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import softeer.wantcar.cartalog.dto.TrimExteriorColorListResponseDto;
-import softeer.wantcar.cartalog.dto.TrimInteriorColorListResponseDto;
+import softeer.wantcar.cartalog.trim.dto.TrimExteriorColorListResponseDto;
+import softeer.wantcar.cartalog.trim.dto.TrimInteriorColorListResponseDto;
 import softeer.wantcar.cartalog.entity.model.ModelExteriorColor;
 import softeer.wantcar.cartalog.entity.model.ModelInteriorColor;
-import softeer.wantcar.cartalog.service.MockTrimColorService;
-import softeer.wantcar.cartalog.service.TrimColorService;
+import softeer.wantcar.cartalog.trim.service.MockTrimColorService;
+import softeer.wantcar.cartalog.trim.service.TrimColorService;
+import softeer.wantcar.cartalog.trim.controller.TrimColorController;
 
 import java.util.List;
