+++ backend/src/test/java/softeer/wantcar/cartalog/controller/TrimControllerTest.java
@@ -6,13 +6,14 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import softeer.wantcar.cartalog.dto.HMGDataDto;
-import softeer.wantcar.cartalog.dto.TrimListResponseDTO;
+import softeer.wantcar.cartalog.global.dto.HMGDataDto;
+import softeer.wantcar.cartalog.trim.dto.TrimListResponseDTO;
+import softeer.wantcar.cartalog.trim.controller.TrimController;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static softeer.wantcar.cartalog.dto.TrimListResponseDTO.*;
+import static softeer.wantcar.cartalog.trim.dto.TrimListResponseDTO.*;
 
 @DisplayName("트림 도메인 컨트롤러 테스트")
 class TrimControllerTest {
